Moving AST relaated staf in separate package. Also moving ASTNodeType enum out of AST.
