+++ aibolit/metrics/RFC/rfc.py
@@ -1,5 +1,5 @@
-from aibolit.utils.ast import AST, ASTNodeType
-from aibolit.utils.java_package import JavaPackage
+from aibolit.ast import AST, ASTNodeType
+from aibolit.ast.java_package import JavaPackage
 from typing import Set, Any
