+++ aibolit/metrics/countLeaves/numberofleaves.py
@@ -1,5 +1,5 @@
 import javalang
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 from typing import List, Type, Any
