+++ aibolit/metrics/fanout/FanOut.py
@@ -1,7 +1,6 @@
-from aibolit.utils.ast_builder import build_ast
-from aibolit.utils.ast import AST, ASTNodeType
+from aibolit.ast import build_ast, AST, ASTNodeType
 from itertools import islice
-from aibolit.utils.java_package import JavaPackage
+from aibolit.ast.java_package import JavaPackage
 
 
 class FanOut:
