+++ aibolit/patterns/joined_validation/joined_validation.py
@@ -21,8 +21,7 @@
 # SOFTWARE.
 
 from typing import List, Tuple
-from aibolit.utils.ast_builder import build_ast
-from aibolit.utils.ast import AST, ASTNodeType
+from aibolit.ast import AST, ASTNodeType, build_ast
 
 
 class JoinedValidation:
