+++ aibolit/patterns/method_chaining/method_chaining.py
@@ -3,7 +3,7 @@ import javalang
 import uuid
 from collections import defaultdict
 
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 class MethodChainFind:
