+++ aibolit/patterns/multiple_try/multiple_try.py
@@ -6,7 +6,7 @@ import hashlib
 import itertools
 from javalang.tree import FormalParameter
 
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 class MultipleTry:
