+++ aibolit/patterns/multiple_while/multiple_while.py
@@ -1,5 +1,4 @@
-from aibolit.utils.ast import AST, ASTNodeType
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import AST, ASTNodeType, build_ast
 
 
 class MultipleWhile:
