+++ aibolit/patterns/non_final_class/non_final_class.py
@@ -2,7 +2,7 @@ from typing import List
 from javalang.tree import ClassDeclaration
 
 from aibolit.types_decl import LineNumber
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 
 class NonFinalClass:
