+++ aibolit/patterns/private_static_method/private_static_method.py
@@ -1,6 +1,6 @@
 import javalang
 
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 
 class PrivateStaticMethod:
