+++ aibolit/patterns/return_null/return_null.py
@@ -21,8 +21,7 @@
 # SOFTWARE.
 
 from typing import List
-from aibolit.utils.ast_builder import build_ast
-from aibolit.utils.ast import AST, ASTNodeType
+from aibolit.ast import AST, ASTNodeType, build_ast
 
 
 class ReturnNull:
