+++ aibolit/patterns/send_null/send_null.py
@@ -1,7 +1,7 @@
 import javalang
 from typing import List
 
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 
 class SendNull:
