+++ aibolit/patterns/this_finder/this_finder.py
@@ -22,7 +22,7 @@
 
 import javalang
 
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 
 
 class ThisFinder:
