+++ aibolit/patterns/var_siblings/var_siblings.py
@@ -1,6 +1,6 @@
 import javalang
 import re
-from aibolit.utils.ast_builder import build_ast
+from aibolit.ast import build_ast
 import itertools
